Move Rows and Statement to own modules
